Fixed the pom.xml name and description tags
added an entity and a service packages

fixed a typo in the getMonthyPayment method
